move get_feature_dims to utils.py
